requirements.txt
fastapi
uvicorn
tensorflow
keras
keras_preprocessing
pillow
numpy